Move repr test to test_interactive.py and add test for _limit to test_sql_compute
